feat: Refactor regional stats to aggregate by state and conditionally display city blackouts, introduce new Supabase migrations for historical stats and cleanup, and remove numerous old utility scripts.

+++ supabase/migrations/stats_rpc.sql
@@ -110,8 +110,8 @@ BEGIN
   )
   SELECT 
     t.ip,
-    t.lat,
-    t.lon,
+    t.lat::FLOAT,
+    t.lon::FLOAT,
     t.provider,
     t.state,
     t.city,
